refactor(agent-tars-cli): decouple 2
