+++ multimodal/agent-tars-cli/src/builtin-agents/agent-tars/agent-tars-cli-core.ts
@@ -1,0 +1,3 @@
+class AgentTARSCLICore {
+  name = 'agent-tars';
+}
