+++ multimodal/agent-tars-cli/src/core/commands/index.ts
